+++ src/main/java/io/github/astrapi69/crypt/api/algorithm/cert/CertPathValidatorAlgorithm.java
@@ -22,50 +22,45 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.crypt.api.type;
+package io.github.astrapi69.crypt.api.algorithm.cert;
 
-import java.security.KeyStore;
+import java.security.cert.CertPathValidator;
+
+import io.github.astrapi69.crypt.api.algorithm.AesAlgorithm;
+import io.github.astrapi69.crypt.api.algorithm.Algorithm;
 
 /**
- * The enum {@link KeystoreType} for types that are used to create {@link KeyStore} objects
+ * The enum {@link CertPathValidatorAlgorithm} provides all the algorithm names that can be
+ * specified when generating an instance of {@link CertPathValidator}. For more info see: <a href=
+ * "https://docs.oracle.com/en/java/javase/11/docs/specs/security/standard-names.html#certpathvalidator-algorithms">
+ * https://docs.oracle.com/en/java/javase/11/docs/specs/security/standard-names.html#certpathvalidator-algorithms</a>
  */
-public enum KeystoreType implements Type
+public enum CertPathValidatorAlgorithm implements Algorithm
 {
-	/** The enum constant for DKS type */
-	DKS("dks"),
 
-	/** The enum constant for JKS type */
-	JKS("jks"),
+	/** The enum constant for PKIX algorithm */
+	PKIX(CertPathBuilderAlgorithm.PKIX_ALGORITHM_NAME);
 
-	/** The enum constant for JCEKS type */
-	JCEKS("jceks"),
-
-	/** The enum constant for PKCS11 type */
-	PKCS11("pkcs11"),
-
-	/** The enum constant for PKCS12 type */
-	PKCS12("pkcs12");
-
-	/** The type. */
-	private final String type;
+	/** The algorithm */
+	private final String algorithm;
 
 	/**
-	 * Instantiates a new {@link KeystoreType} object
+	 * Instantiates a new {@link AesAlgorithm} object
 	 *
-	 * @param type
-	 *            the type
+	 * @param algorithm
+	 *            the algorithm
 	 */
-	KeystoreType(final String type)
+	CertPathValidatorAlgorithm(final String algorithm)
 	{
-		this.type = type;
+		this.algorithm = algorithm;
 	}
 
 	/**
 	 * {@inheritDoc}
 	 */
 	@Override
-	public String getType()
+	public String getAlgorithm()
 	{
-		return type;
+		return algorithm;
 	}
 }