+++ src/main/java/io/github/astrapi69/crypt/api/type/CertStoreType.java
@@ -24,38 +24,39 @@
  */
 package io.github.astrapi69.crypt.api.type;
 
-import java.security.KeyStore;
+import java.security.cert.CertStore;
 
 /**
- * The enum {@link KeystoreType} for types that are used to create {@link KeyStore} objects
+ * The enum {@link CertStoreType} provides all the type names that can be specified when generating
+ * an instance of {@link CertStore}. For more info see: <a href=
+ * "https://docs.oracle.com/en/java/javase/11/docs/specs/security/standard-names.html#certstore-types">
+ * https://docs.oracle.com/en/java/javase/11/docs/specs/security/standard-names.html#certstore-types</a>
  */
-public enum KeystoreType implements Type
+public enum CertStoreType implements Type
 {
-	/** The enum constant for DKS type */
-	DKS("dks"),
 
-	/** The enum constant for JKS type */
-	JKS("jks"),
+	/** The enum constant for Collection type */
+	Collection(CertStoreType.COLLECTION_TYPE_NAME),
 
-	/** The enum constant for JCEKS type */
-	JCEKS("jceks"),
+	/** The enum constant for LDAP type */
+	LDAP(CertStoreType.LDAP_TYPE_NAME);
 
-	/** The enum constant for PKCS11 type */
-	PKCS11("pkcs11"),
+	/** The string constant COLLECTION_TYPE_NAME */
+	public static final String COLLECTION_TYPE_NAME = "Collection";
 
-	/** The enum constant for PKCS12 type */
-	PKCS12("pkcs12");
+	/** The string constant LDAP_TYPE_NAME */
+	public static final String LDAP_TYPE_NAME = "LDAP";
 
 	/** The type. */
 	private final String type;
 
 	/**
-	 * Instantiates a new {@link KeystoreType} object
+	 * Instantiates a new {@link CertStoreType} object
 	 *
 	 * @param type
 	 *            the type
 	 */
-	KeystoreType(final String type)
+	CertStoreType(final String type)
 	{
 		this.type = type;
 	}