+++ src/main/java/io/github/astrapi69/crypt/api/type/CertificateFactoryType.java
@@ -24,38 +24,34 @@
  */
 package io.github.astrapi69.crypt.api.type;
 
-import java.security.KeyStore;
+import java.security.cert.CertificateFactory;
 
 /**
- * The enum {@link KeystoreType} for types that are used to create {@link KeyStore} objects
+ * The enum {@link CertificateFactoryType} provides all the type names that can be specified when
+ * generating an instance of {@link CertificateFactory}. For more info see: <a href=
+ * "https://docs.oracle.com/en/java/javase/11/docs/specs/security/standard-names.html#certificatefactory-types">
+ * https://docs.oracle.com/en/java/javase/11/docs/specs/security/standard-names.html#certificatefactory-types</a>
  */
-public enum KeystoreType implements Type
+public enum CertificateFactoryType implements Type
 {
-	/** The enum constant for DKS type */
-	DKS("dks"),
 
-	/** The enum constant for JKS type */
-	JKS("jks"),
+	/** The enum constant for X.509 certificate type */
+	X_509(CertificateFactoryType.X_509_CERTIFICATE_TYPE_NAME);
 
-	/** The enum constant for JCEKS type */
-	JCEKS("jceks"),
 
-	/** The enum constant for PKCS11 type */
-	PKCS11("pkcs11"),
-
-	/** The enum constant for PKCS12 type */
-	PKCS12("pkcs12");
+	/** The string constant X_509_CERTIFICATE_TYPE_NAME */
+	public static final String X_509_CERTIFICATE_TYPE_NAME = "X.509";
 
 	/** The type. */
 	private final String type;
 
 	/**
-	 * Instantiates a new {@link KeystoreType} object
+	 * Instantiates a new {@link CertificateFactoryType} object
 	 *
 	 * @param type
 	 *            the type
 	 */
-	KeystoreType(final String type)
+	CertificateFactoryType(final String type)
 	{
 		this.type = type;
 	}